Extract compute_mm_prefix_range_tensor into backends/utils.py
Move compute_mm_prefix_range_tensor from a static method on
TritonAttentionMetadata to a standalone utility in backends/utils.py
so both Triton and FlashAttention backends can import it directly
without cross-backend dependencies.

Also guard supports_mm_prefix behind is_fa_version_supported(4).
